Move ImageJMacro and Matlab token makers to highliters package
They are not used outside of that package.
